Updating spring boot to 3.0.2 / Update to java 19 / using records / adding testcontainers with postgres/  removing lombok

+++ poc-hexagonal_architecture/src/main/java/es/rbailen/sample/hexagonalarchitecture/application/ports/output/ProductOutputPort.java
@@ -8,6 +8,6 @@ public interface ProductOutputPort {
 
     Product saveProduct(Product product);
 
-    Optional<Product> getProductById(Long id);
+    Optional<Product> getProductById(Integer id);
 
 }
